Navigation Updates & Basic Tower Defense Mini-Game
Updated the Navigation Logic in the basic mob so they have a smoother jump logic and pathfinding logic. There's still additional conditionals that need to be worked but it's start.

Also added a basic tower defense game with some gui placement and logic and basic upgrade options. Should hopefully pave the path to getting the rest of the core tower defense logic into the practice game